+++ NEWS.md
@@ -1,3 +1,12 @@
+# toastui 0.3.2
+
+### Breaking changes
+
+* `grid_proxy_delete_row()`: argument `index` renamed `rowKey` and now expect the row key of the row to delete, you can find the `rowKey` value in `input$<outputId>_data`.
+* `input$<outputId>_data` is now given with an extra column `rowKey` thats gives the internant row key of the row.
+
+
+
 # toastui 0.3.1
 
 * Updated tui-grid to 4.21.17.